Interchangable agent start command for env difference
For instance, onprem environment needs a different start command.
So far agent start command was hard-coded to the one for linux-ec2 environment.
While doing so, did some refactoring of metadata in order to make this kind of changes easier in the future, as more and more flags have been getting added and it's reasonable to assume that the trend will continue.